Revert "refactor(api): Move opentrons.initialize() to its own file (#15191)"
This reverts commit 1be63cd829baa8d2e02b927cb0ebf33a6bc386c3.
