+++ g-code-testing/g_code_test_data/protocol/protocol_configurations.py
@@ -1,4 +1,4 @@
-from opentrons.protocols.api_support.types import APIVersion
+from opentrons import APIVersion
 
 from opentrons.hardware_control.emulation.settings import (
     Settings,
